// src/types/Calculations.ts
export type SavingsCalculations = {
  solarPanels: number;
  batterySize: number;
  bedrooms: number;
  baselineUtilityBill: number;
  solarOnlyUtilityBill: number;
  solarBatteryUtilityBill: number;
  solarBatteryLumoUtilityBill: number;
  estimatedSystemPrice: number;
  depositPercentage: string;
  yearsFinanced: number;
  interestRate: string;
};